pi_mod_02/requirements.txt
jupyter
pandas
numpy
python-dotenv
psycopg2-binary
python-decouple
sqlalchemy
matplotlib 
seaborn